+++ apps/api/src/entry/dev.ts
@@ -108,6 +108,7 @@ async function start() {
     logger,
     corsOrigin,
     isDev,
+    enableLocalDev: true,
     nodeRegistry,
     remoteNodeRegistry,
     appName,